+++ package.json
@@ -12,11 +12,14 @@
     "@liveblocks/client": "^1.10.2",
     "@liveblocks/node": "^1.10.2",
     "@liveblocks/react": "^1.10.2",
+    "clsx": "^2.1.0",
     "nanoid": "^5.0.6",
     "next": "14.1.3",
     "perfect-freehand": "^1.2.2",
     "react": "^18",
+    "react-contenteditable": "^3.3.7",
     "react-dom": "^18",
+    "tailwind-merge": "^2.2.2",
     "zustand": "^4.5.2"
   },
   "devDependencies": {
