+++ src/app/page.tsx
@@ -1,9 +1,3 @@
-import Room from "@/components/Room";
-
 export default function Home() {
-  return (
-    <main>
-      <Room />
-    </main>
-  );
+  return <main></main>;
 }
