+++ src/lib/types.ts
@@ -8,6 +8,8 @@ export enum LayerType {
   Rectangle,
   Ellipse,
   Path,
+  Text,
+  Note,
 }
 
 export type Camera = {
@@ -15,7 +17,7 @@ export type Camera = {
   y: number;
 };
 
-export type Layer = RectangleLayer | EllipseLayer | PathLayer;
+export type Layer = RectangleLayer | EllipseLayer | PathLayer | TextLayer;
 
 export type RectangleLayer = {
   type: LayerType.Rectangle;
@@ -45,6 +47,16 @@ export type PathLayer = {
   width: number;
   fill: Color;
   points: number[][];
+};
+
+export type TextLayer = {
+  type: LayerType.Text;
+  x: number;
+  y: number;
+  height: number;
+  width: number;
+  fill: Color;
+  value?: string;
 };
 
 export type Point = {
@@ -81,7 +93,7 @@ export type CanvasState =
     }
   | {
       mode: CanvasMode.Inserting;
-      layerType: LayerType.Ellipse | LayerType.Rectangle;
+      layerType: LayerType.Ellipse | LayerType.Rectangle | LayerType.Text | LayerType.Note;
     }
   | {
       mode: CanvasMode.Pencil;
