+++ src/lib/utils.ts
@@ -1,3 +1,4 @@
+import { type ClassValue, clsx } from "clsx";
 import {
   Color,
   Side,
@@ -9,6 +10,11 @@ import {
   PathLayer,
   Camera,
 } from "./types";
+import { twMerge } from "tailwind-merge";
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
 
 export function colorToCss(color: Color) {
   return `#${color.r.toString(16).padStart(2, "0")}${color.g
